+++ components/Icons/Frameworks.jsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { useEffect, useState } from "react";
 
 // CSS
