+++ components/Icons/LanguageFaces.jsx
@@ -1,3 +1,5 @@
+"use client";
+
 import languages from "../public/Assets/languages.json";
 
 var projectsJSON = JSON.parse(JSON.stringify(languages));
