+++ components/NavBar/NavBar.jsx
@@ -1,9 +1,11 @@
+"use client";
+
 import React, { useEffect, useState } from "react";
 import { useTheme } from "next-themes";
 import Link from "next/link";
 
 // Child Components
-import PixelSwitch from "../../components/PixelSwitch/PixelSwitch";
+import PixelSwitch from "@components/PixelSwitch/PixelSwitch";
 
 // CSS
 import NavBarCSS from "./NavBar.module.css";
