+++ components/Projects/Projects.jsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { useEffect, useState } from "react";
 import { useTheme } from "next-themes";
 
@@ -46,7 +48,9 @@ export default function Projects(props) {
     return (
         <div>
             <div
-                className={`${utilStyles.imageRendering} ${ProjectsCSS.projectsBacker} ${background()} `}
+                className={`${utilStyles.imageRendering} ${
+                    ProjectsCSS.projectsBacker
+                } ${background()} `}
             >
                 <section className={ProjectsCSS.projectTitle}>
                     <p>{props.name}</p>
