+++ hooks/MainButtons.jsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useTheme } from "next-themes";
 
 export function returnURL(text) {
