+++ jsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "paths": {
+      "@app/*": ["/app/*"],
+      "@components/*": ["/components/*"]
+    }
+  }
+}
