+++ package.json
@@ -7,12 +7,12 @@
   },
   "dependencies": {
     "js-pixel-fonts": "^1.6.0",
-    "next": "^14.1.4",
+    "next": "^15.3.1",
     "next-themes": "^0.2.1",
     "node-html-parser": "^6.1.12",
-    "react": "18.2.0",
+    "react": "^19.1.0",
     "react-circular-progressbar": "^2.1.0",
-    "react-dom": "18.2.0",
+    "react-dom": "^19.1.0",
     "react-grid-gallery": "^1.0.0",
     "react-image-gallery": "^1.3.0",
     "react-responsive-carousel": "^3.2.23",
